+++ lua/selection_command.lua
@@ -4,6 +4,7 @@ local C = require'context'
 local L = require'tools.list'
 local complete = require'selection.completer'.complete
 local api = require'nvimapi'
+local replace_lines = require'selection_command.replace_lines'
 
 local function execute(...)
   local ctxt  = C.context()
@@ -17,27 +18,10 @@ local function execute(...)
   -- end
 end
 
-local function rubocop_command(lines, cop)
-  local prefix = string.gsub(lines[1], '[^%s].*$', '')
-  return L.flatten(
-    prefix .. "# rubocop:disable " .. cop,
-    lines,
-    prefix .. "# rubocop:enable " .. cop)
-end
-
-local defined_lines_commands = {
-  rubocop = rubocop_command,
-}
-
-local function execute_lines_command(lines, command, ...)
-  local command = defined_lines_commands[command]
-  if not command then return lines end
-  return command(lines, ...)
-end
-
 local function execute_lines(...)
   local fl, ll, lines = api.get_selected_lines()
-  local result = execute_lines_command(lines, ...)
+
+  local result = replace_lines(lines, ...)
   api.set_lines(fl, ll, result)
 end